emobility-echarging: do not do concurrent sync jobs
Some of these sync scheduled jobs take longer then their re-scheduling,
so the former call fails or overwrites data that has been half-synced.
To avoid this we use a single sync call, and put just a single job per
scheduler. That is the job will not trigger, if it is still running.